week6-agent-implementation/requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
openai>=1.10.0
requests>=2.31.0
python-dotenv>=1.0.0
